+++ plugins/plugin-itch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/plugin-itch
 
+## 1.0.1-latest.24
+
+### Patch Changes
+
+- sqd
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.1-latest.29
+
 ## 1.0.1-beta.23
 
 ### Patch Changes